bugfix/correcting_paths to pages (#34)
